// code-app/src/generated/services/Dpa_invoicesService.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */

import type { Dpa_invoicesBase, Dpa_invoices } from '../models/Dpa_invoicesModel';
import type { Dpa_invoicesUploadColumnName } from '../models/Dpa_invoicesModel';
import type { GetEntityMetadataOptions, EntityMetadata } from '@microsoft/power-apps/data/metadata/dataverse';
import type { IGetOptions, IGetAllOptions } from '../models/CommonModels';
import type { IOperationResult } from '@microsoft/power-apps/data';
import { dataSourcesInfo } from '../../../.power/schemas/appschemas/dataSourcesInfo';
import { getClient } from '@microsoft/power-apps/data';


export class Dpa_invoicesService {
  private static readonly dataSourceName = 'dpa_invoices';

  private static readonly client = getClient(dataSourcesInfo);

  public static async create(record: Omit<Dpa_invoicesBase, 'dpa_invoiceid'>): Promise<IOperationResult<Dpa_invoices>> {
    const result = await Dpa_invoicesService.client.createRecordAsync<Omit<Dpa_invoicesBase, 'dpa_invoiceid'>, Dpa_invoices>(
      Dpa_invoicesService.dataSourceName,
      record
    );
    return result;
  }

  public static async update(id: string, changedFields: Partial<Omit<Dpa_invoicesBase, 'dpa_invoiceid'>>): Promise<IOperationResult<Dpa_invoices>> {
    const result = await Dpa_invoicesService.client.updateRecordAsync<Partial<Omit<Dpa_invoicesBase, 'dpa_invoiceid'>>, Dpa_invoices>(
      Dpa_invoicesService.dataSourceName,
      id,
      changedFields
    );
    return result;
  }

  public static async delete(id: string): Promise<void> {
    await Dpa_invoicesService.client.deleteRecordAsync(
      Dpa_invoicesService.dataSourceName,
      id);
  }

  public static async get(id: string, options?: IGetOptions): Promise<IOperationResult<Dpa_invoices>> {
    const result = await Dpa_invoicesService.client.retrieveRecordAsync<Dpa_invoices>(
      Dpa_invoicesService.dataSourceName,
      id,
      options
    );
    return result;
  }

  public static async getAll(options?: IGetAllOptions): Promise<IOperationResult<Dpa_invoices[]>> {
    const result = await Dpa_invoicesService.client.retrieveMultipleRecordsAsync<Dpa_invoices>(
      Dpa_invoicesService.dataSourceName,
      options
    );
    return result;
  }

  public static getMetadata(
    options: GetEntityMetadataOptions<Dpa_invoices> = {}
  ): Promise<IOperationResult<Partial<EntityMetadata>>> {
    return Dpa_invoicesService.client.executeAsync({
      dataverseRequest: {
        action: "getEntityMetadata",
        parameters: {
          tableName: Dpa_invoicesService.dataSourceName,
          options: options as GetEntityMetadataOptions,
        },
      },
    });
  }

  public static async upload(id: string, columnName: Dpa_invoicesUploadColumnName, file: File, fileDisplayName?: string): Promise<IOperationResult<void>> {
    const arrayBuffer = await file.arrayBuffer();
    const data = new Uint8Array(arrayBuffer);
    const result = await Dpa_invoicesService.client.uploadFileToRecord(
      Dpa_invoicesService.dataSourceName,
      id,
      columnName,
      fileDisplayName || file.name,
      data,
    );
    return result;
  }
}