// code-app/src/generated/services/Dpa_receiptsService.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */

import type { Dpa_receiptsBase, Dpa_receipts } from '../models/Dpa_receiptsModel';
import type { Dpa_receiptsUploadColumnName } from '../models/Dpa_receiptsModel';
import type { GetEntityMetadataOptions, EntityMetadata } from '@microsoft/power-apps/data/metadata/dataverse';
import type { IGetOptions, IGetAllOptions } from '../models/CommonModels';
import type { IOperationResult } from '@microsoft/power-apps/data';
import { dataSourcesInfo } from '../../../.power/schemas/appschemas/dataSourcesInfo';
import { getClient } from '@microsoft/power-apps/data';


export class Dpa_receiptsService {
  private static readonly dataSourceName = 'dpa_receipts';

  private static readonly client = getClient(dataSourcesInfo);

  public static async create(record: Omit<Dpa_receiptsBase, 'dpa_receiptid'>): Promise<IOperationResult<Dpa_receipts>> {
    const result = await Dpa_receiptsService.client.createRecordAsync<Omit<Dpa_receiptsBase, 'dpa_receiptid'>, Dpa_receipts>(
      Dpa_receiptsService.dataSourceName,
      record
    );
    return result;
  }

  public static async update(id: string, changedFields: Partial<Omit<Dpa_receiptsBase, 'dpa_receiptid'>>): Promise<IOperationResult<Dpa_receipts>> {
    const result = await Dpa_receiptsService.client.updateRecordAsync<Partial<Omit<Dpa_receiptsBase, 'dpa_receiptid'>>, Dpa_receipts>(
      Dpa_receiptsService.dataSourceName,
      id,
      changedFields
    );
    return result;
  }

  public static async delete(id: string): Promise<void> {
    await Dpa_receiptsService.client.deleteRecordAsync(
      Dpa_receiptsService.dataSourceName,
      id);
  }

  public static async get(id: string, options?: IGetOptions): Promise<IOperationResult<Dpa_receipts>> {
    const result = await Dpa_receiptsService.client.retrieveRecordAsync<Dpa_receipts>(
      Dpa_receiptsService.dataSourceName,
      id,
      options
    );
    return result;
  }

  public static async getAll(options?: IGetAllOptions): Promise<IOperationResult<Dpa_receipts[]>> {
    const result = await Dpa_receiptsService.client.retrieveMultipleRecordsAsync<Dpa_receipts>(
      Dpa_receiptsService.dataSourceName,
      options
    );
    return result;
  }

  public static getMetadata(
    options: GetEntityMetadataOptions<Dpa_receipts> = {}
  ): Promise<IOperationResult<Partial<EntityMetadata>>> {
    return Dpa_receiptsService.client.executeAsync({
      dataverseRequest: {
        action: "getEntityMetadata",
        parameters: {
          tableName: Dpa_receiptsService.dataSourceName,
          options: options as GetEntityMetadataOptions,
        },
      },
    });
  }

  public static async upload(id: string, columnName: Dpa_receiptsUploadColumnName, file: File, fileDisplayName?: string): Promise<IOperationResult<void>> {
    const arrayBuffer = await file.arrayBuffer();
    const data = new Uint8Array(arrayBuffer);
    const result = await Dpa_receiptsService.client.uploadFileToRecord(
      Dpa_receiptsService.dataSourceName,
      id,
      columnName,
      fileDisplayName || file.name,
      data,
    );
    return result;
  }
}