// code-app/src/generated/models/Dpa_invoicesModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Dpa_invoicesdpa_extractionmethod = {
  100000000: 'NativeAIBuilderModel',
  100000001: 'AIBuilderCustomPrompt'
} as const;
export type Dpa_invoicesdpa_extractionmethod = keyof typeof Dpa_invoicesdpa_extractionmethod;
export const Dpa_invoicesdpa_processingstatus = {
  100000000: 'Draft',
  100000001: 'ReadytoProcess',
  100000002: 'Processing',
  100000003: 'Completed',
  100000004: 'Failed'
} as const;
export type Dpa_invoicesdpa_processingstatus = keyof typeof Dpa_invoicesdpa_processingstatus;
export const Dpa_invoicesstatecode = {
  0: 'Active',
  1: 'Inactive'
} as const;
export type Dpa_invoicesstatecode = keyof typeof Dpa_invoicesstatecode;
export const Dpa_invoicesstatuscode = {
  1: 'Active',
  2: 'Inactive'
} as const;
export type Dpa_invoicesstatuscode = keyof typeof Dpa_invoicesstatuscode;

export interface Dpa_invoicesBase {
  dpa_amountdue?: number;
  dpa_billtoaddress?: string;
  dpa_billtoname?: string;
  dpa_currencycode?: string;
  dpa_duedate?: string;
  dpa_extractedlineitems?: string;
  dpa_extractionconfidence?: number;
  dpa_extractionmethod: Dpa_invoicesdpa_extractionmethod;
  dpa_invoicedate?: string;
  dpa_invoiceid: string;
  dpa_invoicename?: string;
  dpa_invoicenumber?: string;
  dpa_lastprocessingrunid?: string;
  dpa_processedby?: string;
  dpa_processingattemptcount?: number;
  dpa_processingcompletedon?: string;
  dpa_processingerrorcode?: string;
  dpa_processingerrormessage?: string;
  dpa_processingstartedon?: string;
  dpa_processingstatus: Dpa_invoicesdpa_processingstatus;
  dpa_purchaseordernumber?: string;
  dpa_rawextractionresult?: string;
  dpa_reprocessrequested?: boolean;
  dpa_subtotal?: number;
  dpa_taxamount?: number;
  dpa_totalamount?: number;
  dpa_vendoraddress?: string;
  dpa_vendorname?: string;
  dpa_vendortaxnumber?: string;
  importsequencenumber?: number;
  overriddencreatedon?: string;
  ownerid: string;
  owneridtype: string;
  statecode: Dpa_invoicesstatecode;
  statuscode?: Dpa_invoicesstatuscode;
  timezoneruleversionnumber?: number;
  "TransactionCurrencyId@odata.bind"?: string;
  utcconversiontimezonecode?: number;
}

export interface Dpa_invoices extends Dpa_invoicesBase {
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  dpa_amountdue_base?: number;
  dpa_extractionmethodname?: string;
  dpa_processingstatusname?: string;
  dpa_reprocessrequestedname?: string;
  dpa_sourcedocument?: string;
  dpa_sourcedocument_name?: string;
  dpa_subtotal_base?: number;
  dpa_taxamount_base?: number;
  dpa_totalamount_base?: number;
  exchangerate?: number;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  owneridname: string;
  owneridyominame: string;
  owningbusinessunitname: string;
  statecodename?: string;
  statuscodename?: string;
  transactioncurrencyidname?: string;
  versionnumber?: number;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  owningbusinessunit?: object;
  _owningbusinessunit_value?: string;
  owningteam?: object;
  _owningteam_value?: string;
  owninguser?: object;
  _owninguser_value?: string;
  transactioncurrencyid?: object;
  _transactioncurrencyid_value?: string;
}
export type Dpa_invoicesUploadColumnName = 'dpa_sourcedocument';